+++ UnityProject/Assets/Scripts/Weapons/Projectiles/BulletBehaviour.cs
@@ -14,7 +14,7 @@ public abstract class BulletBehaviour : MonoBehaviour
 {
 	private BodyPartType bodyAim;
 	[Range(0, 100)]
-	public int damage = 25;
+	public float damage = 25;
 	private GameObject shooter;
 	private Weapon weapon;
 	public DamageType damageType;
